+++ lib/kettle/family/workflow.rb
@@ -10,17 +10,21 @@ module Kettle
         "docs" => "bundle exec rake yard"
       }.freeze
 
-      def initialize(command:, config:, members:, execute: false, commit: false, allow_dirty: false)
+      def initialize(command:, config:, members:, execute: false, commit: false, allow_dirty: false, publish: false, push: false, tag: false)
         @command = command
         @config = config
         @members = members
         @execute = execute
         @commit = commit
         @allow_dirty = allow_dirty
+        @publish = publish
+        @push = push
+        @tag = tag
       end
 
       def results
         return check_results if command == "check"
+        return release_results if command == "release"
         guard_family_commit!
 
         runner = CommandRunner.new(execute: execute)
@@ -38,10 +42,44 @@ module Kettle
 
       private
 
-      attr_reader :command, :config, :members, :execute, :commit, :allow_dirty
+      attr_reader :command, :config, :members, :execute, :commit, :allow_dirty, :publish, :push, :tag
 
       def check_results
         members.map { |member| ReadinessCheck.call(member: member) }
+      end
+
+      def release_results
+        runner = CommandRunner.new(execute: execute)
+        members.each_with_object([]) do |member, memo|
+          append_release_internal_checks(member: member, memo: memo)
+          break memo unless memo.last(2).all?(&:ok?)
+
+          memo << runner.call(member: member, phase: release_phase, command: release_command)
+          break memo unless memo.last.ok?
+
+          append_release_git_phases(member: member, runner: runner, memo: memo)
+          break memo unless memo.last.ok?
+        end
+      end
+
+      def append_release_internal_checks(member:, memo:)
+        memo << ReadinessCheck.call(member: member)
+        memo << ChangelogCheck.call(member: member) if memo.last.ok?
+      end
+
+      def release_phase
+        publish ? "release_publish" : "release_build"
+      end
+
+      def release_command
+        publish ? config.release_publish_command : config.release_build_command
+      end
+
+      def append_release_git_phases(member:, runner:, memo:)
+        memo << runner.call(member: member, phase: "release_tag", command: config.release_tag_command) if tag
+        return if memo.any? && !memo.last.ok?
+
+        memo << runner.call(member: member, phase: "release_push", command: config.release_push_command) if push
       end
 
       def guard_family_commit!